bugfix: background thread crashes when there are no explicit change listeners.
This bug was introduced when I added support for explicitly waiting on account updates. When nothing was awaiting changes and the listeners count was zero, the broadcast channel was erroring out in an attempt to send an update, that caused the worker thread to creash and stop updating the shadowed accounts values.

This fix makes sure that we have at least one listener before broadcasting updates to account changes. Eitherway, background updates to the underlying storage still happen correctly.